requirements.txt
Flask==0.10.1
Flask-RESTful==0.3.0
Flask-Bootstrap==3.3.0.1
Jinja2==2.7.3
pyserial==2.7